requirements.txt
asgiref==3.3.4
Django==3.2.1
django-adminlte2-templates==1.4.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-gravatar2==1.4.4
django-jalali-date==0.3.1
django-utils-six==2.0
jdatetime==3.6.2
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
